Reposition report button and new form query
