+++ components/contact-details.tsx
@@ -3,7 +3,6 @@ import { anchorTransformer } from "../lib/htmr-transformers";
 import { Contact } from "../lib/provinces";
 import { isNotEmpty, stripTags } from "../lib/string-utils";
 
-import { ExclamationCircleIcon } from "@heroicons/react/outline";
 import htmr from "htmr";
 import { HtmrOptions } from "htmr/src/types";
 
@@ -12,47 +11,10 @@ type ContactDetailsProps = {
   provinceName: string;
 };
 
-type ReportButtonProps = {
-  onClick: () => void;
-};
-
-type ReportButtonHandlerProps = {
-  contact: Contact;
-  provinceName: string;
-};
-
-const reportButtonHandler = (props: ReportButtonHandlerProps) => {
-  const {
-    provinceName,
-    contact: { keterangan, penyedia, kontak },
-  } = props;
-
-  const formBaseUrl =
-    "https://docs.google.com/forms/d/e/1FAIpQLSeDdZLMD9RTfOiKMItNb8542KAO5w3x9O2ZFQmNyff1rYRZvQ/viewform?usp=pp_url";
-
-  const formQuery = `&entry.346789668=__other_option__&entry.346789668.other_option_response=${keterangan}&entry.323081545=${provinceName}&entry.68818336=${penyedia}&entry.217416134=${kontak}`;
-
-  return window.open(formBaseUrl.concat(encodeURI(formQuery)), "_blank");
-};
-
-const ReportButton = (props: ReportButtonProps) => (
-  <span className="ml-4 flex-shrink-0">
-    <button
-      className="bg-white rounded-md font-medium text-indigo-600 hover:text-indigo-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-      onClick={props.onClick}
-      title="Laporkan kesalahan"
-      type="button"
-    >
-      <ExclamationCircleIcon aria-hidden="true" className="w-5 h-5" />
-    </button>
-  </span>
-);
-
 type DescriptionItemProps = {
   label: string;
   value?: string;
   withCopyButton?: boolean;
-  onClick: () => void;
 };
 
 const DescriptionItem = (props: DescriptionItemProps) => {
@@ -71,65 +33,31 @@ const DescriptionItem = (props: DescriptionItemProps) => {
         {typeof value == "string" &&
           value.length > 0 &&
           props.withCopyButton && <CopyButton text={stripTags(value)} />}
-        <ReportButton onClick={props.onClick} />
       </dd>
     </div>
   );
 };
 
-export function ContactDetails({ contact, provinceName }: ContactDetailsProps) {
-  const _reportButtonHandler = () => {
-    reportButtonHandler({ contact, provinceName });
-  };
-
+export function ContactDetails({ contact }: ContactDetailsProps) {
   return (
     <div className="bg-white shadow overflow-hidden rounded-md">
       <dl className="divide-y divide-gray-200">
-        <DescriptionItem
-          label="Penyedia"
-          onClick={_reportButtonHandler}
-          value={contact.penyedia}
-        />
-        <DescriptionItem
-          label="Keterangan"
-          onClick={_reportButtonHandler}
-          value={contact.keterangan}
-        />
-        <DescriptionItem
-          label="Lokasi"
-          onClick={_reportButtonHandler}
-          value={contact.lokasi}
-        />
-        <DescriptionItem
-          label="Kontak"
-          onClick={_reportButtonHandler}
-          value={contact.kontak}
-          withCopyButton
-        />
-        <DescriptionItem
-          label="Alamat"
-          onClick={_reportButtonHandler}
-          value={contact.alamat}
-          withCopyButton
-        />
-        <DescriptionItem
-          label="Tautan"
-          onClick={_reportButtonHandler}
-          value={contact.link}
-        />
+        <DescriptionItem label="Penyedia" value={contact.penyedia} />
+        <DescriptionItem label="Keterangan" value={contact.keterangan} />
+        <DescriptionItem label="Lokasi" value={contact.lokasi} />
+        <DescriptionItem label="Kontak" value={contact.kontak} />
+        <DescriptionItem label="Alamat" value={contact.alamat} />
+        <DescriptionItem label="Tautan" value={contact.link} />
         <DescriptionItem
           label="Terakhir Update"
-          onClick={_reportButtonHandler}
           value={contact.terakhir_update}
         />
         <DescriptionItem
           label="Bentuk Verifikasi"
-          onClick={_reportButtonHandler}
           value={contact.bentuk_verifikasi}
         />
         <DescriptionItem
           label="Tambahan Informasi"
-          onClick={_reportButtonHandler}
           value={contact.tambahan_informasi}
         />
       </dl>
